refactor: extract data classes from UserProfileApiClient into separate files
  - Move TopicBasedProfileUpdateRequest to dedicated file
  - Move EventBasedProfileUpdateRequest to dedicated file
  - Move UserProfileUpdateResponse to dedicated file
  - Clean up UserProfileApiClient by removing embedded data classes
