+++ src/main/kotlin/com/flownews/api/interaction/infra/UserProfileUpdateResponse.kt
@@ -1,0 +1,12 @@
+package com.flownews.api.interaction.infra
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class UserProfileUpdateResponse(
+    @JsonProperty("status")
+    val status: String,
+    @JsonProperty("user_id")
+    val userId: Int,
+    @JsonProperty("processed_count")
+    val processedCount: Int,
+)
